requirements.txt
# Core
streamlit>=1.34.0,<2.0.0
torch==2.5.1
torchvision==0.20.1
timm>=1.0.3,<1.1.0
numpy>=1.26.0,<3.0.0
pillow>=10.2.0

# EXIF + Maps
exifread>=3.0.0
folium>=0.16.0
streamlit-folium>=0.18.0

# Nice-to-haves (used implicitly or common in your env)
tqdm>=4.66.0